remvoe global variable gpt
