+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/alkshmir/go-openai-discord
 
-go 1.20
+go 1.22
 
 require (
 	github.com/bwmarrin/discordgo v0.28.1
